Backend: Auto Regenerate Apple Secret
