+++ .env.template
@@ -74,6 +74,10 @@ SYNC_FOR_CUSTOMER="Test" # Options: "Test", "Osnabrück", "Hannover"
 # Read SSO_APPLE.md
 #AUTH_APPLE_CLIENT_ID=
 #AUTH_APPLE_CLIENT_SECRET=
+# The following secrets will automatically regenerate the Apple Client Secret, in order to prevent expiration after 6 months. This way the Apple Client Secret can be generated automatically and rotated every 3 months.
+#AUTH_APPLE_HOOK_APPLE_TEAM_ID=
+#AUTH_APPLE_HOOK_APPLE_KEY_ID=
+#AUTH_APPLE_HOOK_APPLE_PRIVATE_KEY=
 
 #AUTH_GOOGLE_CLIENT_ID=
 #AUTH_GOOGLE_CLIENT_SECRET=
