+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/package.json
@@ -36,7 +36,7 @@
       {
         "type": "hook",
         "name": "apple-secret-rotator",
-        "source": "src/extensions/hooks/apple-secret-rotator/index.ts"
+        "source": "src/apple-secret-rotator/index.ts"
       },
       {
         "type": "hook",
