+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/src/forms-sync-hook/customers/hannover/FormSyncHannover.ts
@@ -3,6 +3,7 @@ import { MyDatabaseHelper } from '../../../helpers/MyDatabaseHelper';
 import { WorkflowScheduleHelper } from '../../../workflows-runs-hook';
 import { FormHousingContractsWorkflowHannover } from './FormHousingContractsWorkflowHannover';
 import { EnvVariableHelper } from '../../../helpers/EnvVariableHelper';
+import {CronHelper} from "../../../helpers/CronHelper";
 
 const DIRECTUS_TL1_HOUSING_EXPORT_PATH = '/directus/tl1/housingContracts.csv'; // This is defined in docker-compose.yaml statically
 
@@ -21,7 +22,7 @@ export class FormSyncHannover {
       workflowRunInterface: new FormHousingContractsWorkflowHannover(DIRECTUS_TL1_HOUSING_EXPORT_PATH),
       myDatabaseHelper: myDatabaseHelper,
       schedule: triggerContext.schedule,
-      cronOject: WorkflowScheduleHelper.EVERY_HOUR,
+      cronOject: CronHelper.EVERY_HOUR,
     });
   }
 }
