+++ apps/backend/Backend/directusExtensions/directus-extension-rocket-meals-bundle/src/housing-sync-hook/index.ts
@@ -9,6 +9,7 @@ import { SingleWorkflowRun } from '../workflows-runs-hook/WorkflowRunJobInterfac
 import { WorkflowRunContext } from '../helpers/WorkflowRunContext';
 import { DatabaseTypes } from 'repo-depkit-common';
 import { WORKFLOW_RUN_STATE } from '../helpers/itemServiceHelpers/WorkflowsRunEnum';
+import {CronHelper} from "../helpers/CronHelper";
 
 class HousingSyncWorkflow extends SingleWorkflowRun {
   private readonly parserInterface: ApartmentParserInterface;
@@ -59,6 +60,6 @@ export default defineHook(async ({ action, init, schedule }, apiContext) => {
     workflowRunInterface: new HousingSyncWorkflow(usedParser),
     myDatabaseHelper: myDatabaseHelper,
     schedule: schedule,
-    cronOject: WorkflowScheduleHelper.EVERY_DAY_AT_4AM,
+    cronOject: CronHelper.EVERY_DAY_AT_4AM,
   });
 });
