+++ apps/backend/docker-compose.yaml
@@ -141,6 +141,10 @@ services:
       AUTH_APPLE_SCOPE: "name email" # We sadly only get the email and name on the first login - https://developer.apple.com/documentation/sign_in_with_apple/sign_in_with_apple_js/configuring_your_webpage_for_sign_in_with_apple
       AUTH_APPLE_IDENTIFIER_KEY: "sub" # The key in the response that identifies the user
       AUTH_APPLE_PARAMS: '{"response_mode":"form_post"}'
+      # Auto generate and refresh the client secret
+      AUTH_APPLE_HOOK_APPLE_TEAM_ID: "${AUTH_APPLE_HOOK_APPLE_TEAM_ID}"
+      AUTH_APPLE_HOOK_APPLE_KEY_ID: "${AUTH_APPLE_HOOK_APPLE_KEY_ID}"
+      AUTH_APPLE_HOOK_PRIVATE_KEY: "${AUTH_APPLE_HOOK_PRIVATE_KEY}"
 
       # Directus Auth configuration - Google predefined values
       AUTH_GOOGLE_DRIVER: "openid"
