+++ packages/common-backend/index.ts
@@ -1,2 +1,1 @@
 export * from './src/CollectionHelper';
-export * from './src/apple/generateAppleClientSecret';
